Handle value as strings
The value for the `DateInput` component is now always a string. The logic within the component separates it out into the constituent parts and renders the fields appropriately, subsequently handling changes to those fields and firing the `onChange` event.

Updated the tests to reflect the change to the `value` type, along with the reduced number of `onChange` firings now taking place - it will no longer fire if the entered value is unchanged.

Finally, adjusted the Storybook stories as appropriate too.